+++ CHANGES.rst
@@ -1,7 +1,7 @@
 History
 =======
 
-1.2 (unreleased)
+1.2 (2022-12-05)
 ----------------
 
 - Implement ``expires`` and ``expired`` properties on